+++ Frontend/src/lib/components/homepage/Slogan.svelte
@@ -34,7 +34,7 @@
 
 <style lang="postcss">
   #background {
-    @apply animate-scroll absolute inset-0 -z-10 h-full w-full bg-cover bg-center bg-repeat-x;
+    @apply animate-scroll absolute inset-0 -z-10 h-full min-h-[50rem] w-full bg-cover bg-center bg-repeat-x;
   }
 
   #underline-text {