+++ Frontend/src/routes/+page.svelte
@@ -129,7 +129,7 @@
 
 <style lang="postcss">
   section {
-    @apply h-fit w-full transition-all duration-[1500ms] ease-in-out last:mb-0;
+    @apply h-fit min-h-[50rem] w-full transition-all duration-[1500ms] ease-in-out last:mb-0;
     @apply flex flex-col items-center justify-center gap-6 sm:py-16 2xl:py-32;
 
     &:nth-child(2),