+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.07] - 2024-11-09
+
+### Added
+
+- Add contact page
+- Add SelectMenu component
+- Add FAQ answer
+
+### Changed
+
+- Change homepage responsiveness
+
 ## [0.1.06] - 2024-11-08
 
 ### Added